requirements.txt
allure-pytest==2.13.5
allure_python_commons==2.13.5
pytest==8.2.2
Requests==2.32.3
